requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.12.0
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.26.0 